Implement an up-to-date list of MySQL functions in mysql.pas. Used for - popupmenu in query memo - query helpers - SynCompletionProposal Drops function.txt and parser for that file.
Pros:
- No dependency on this file anymore
- Much easier and safer coding
- File-parsing no longer needed

Notes:

The old function.txt was meant to be editable and extensible for users but I never saw or heard anyone who does that, most likely because it wasn't documented. Another problem was that the file resided in the programs-directory, where only administrators have write-access, so in these cases it couldn't be edited anyway.

The PHP file in /extras can be used for creating an up-to-date list for mysql.pas .

As not all functions of the new list are available in all mysql-versions, it would be the best if we would fetch the list by the use of HELP commands. Contra: Would result in poor performancing popupmenu and many servers don't have the help-tables filled.

git-svn-id: http://heidisql.googlecode.com/svn/trunk@716 560a3889-bb33-0410-9f33-43a8a102dc09