Accidently overseen part of previous change/commit.
Btw: The red "busy" and green "ready" icon in the statusbar had been removed now, as that was no longer working with the PNGImageList. Refixes the everlasting bug #1472803. Maybe these two icons can be added sometime, as that can be very informative to the user, but then please with a less hackish solution than previously.

git-svn-id: http://heidisql.googlecode.com/svn/trunk@1235 560a3889-bb33-0410-9f33-43a8a102dc09